requirements.txt
Django==4.2.9
python-decouple==3.8
Pillow==10.2.0
psycopg2-binary==2.9.9
django-crispy-forms==2.1
crispy-bootstrap5==0.7
channels==4.0.0
channels-redis==4.1.0
djangorestframework==3.14.0
django-axes==6.1.1
django-debug-toolbar==4.2.0
black==24.3.0
flake8==7.0.0
pytest==8.1.1
pytest-django==4.7.0
factory-boy==3.3.0
